requirements.txt
anthropic
langgraph
yt-dlp
duckduckgo-search
Pillow
requests
python-dotenv
replicate
openai
